pyproject.toml
[project]
name = "rhoai-ai-feature-sizing"
version = "0.1.0"
description = "AI-powered feature sizing for Jira using LLMs and MCP."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "llama-stack>=0.2.10",
    "llama-stack-client>=0.2.10",
    "requests>=2.32.3",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "llama-index-core>=0.11.0",
    "llama-index-readers-github>=0.2.0",
    "llama-index-readers-web>=0.2.0",
    "llama-index-readers-file>=0.2.0",
    "nest-asyncio>=1.5.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["rhoai_ai_feature_sizing*"]
